Add built artifacts to kit
When the react-app starts, if there is a JSON file missing, it throws an error but *does not automatically reload if the file is created*. We need to start with each JSON already present. This commit adds the built artifacts.

The alternative is to add an instruction to the user to `compile` (truffle or zos) before running `npm run start` on the client folder.