Load contract artifacts from json files
This commit is a hotfix to an issue in the hot loader (see https://github.com/zeppelinos/zeppelin-solidity-hot-loader/pull/5) that causes the contract to not be reloaded under certain circumstances. To bypass it, it simply loads the contract from the JSON artifact instead of from the Solidity source file, so webpack tracks the changes to the JSON file directly.